Move DatabaseHasTables
