+++ README.md
@@ -40,7 +40,7 @@ return [
      * The restore command will fail, if any health checks fail.
      */
     'health-checks' => [
-        \Wnx\LaravelBackupRestore\HealthChecks\DatabaseHasTables::class,
+        \Wnx\LaravelBackupRestore\HealthChecks\Checks\DatabaseHasTables::class,
     ],
 ];
 ```
@@ -129,7 +129,7 @@ Add your health check to the `health-checks`-array in the `config/laravel-backup
 
 ```php
     'health-checks' => [
-        \Wnx\LaravelBackupRestore\HealthChecks\DatabaseHasTables::class,
+        \Wnx\LaravelBackupRestore\HealthChecks\Checks\DatabaseHasTables::class,
         \App\HealthChecks\MyCustomHealthCheck::class,
     ],
 ```
